% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kmer.r
\name{countNMers}
\alias{countNMers}
\title{Count N Mers
Find the weighted value of each nmer}
\usage{
countNMers(fastaFile, weights, n = 6)
}
\arguments{
\item{fastaFile}{The fasta file to be analyzed, each region must be represented in a single line}

\item{weights}{a vector of weight the same length as the number of members in the fasta file}

\item{n}{the number of ks}
}
\description{
Count N Mers
Find the weighted value of each nmer
}
\examples{
library(mulcal)
library(kmerhmm)
fastaFile<-"~/Dropbox/UTX-Alex/jan/combined.fasta"
heightFile<-"~/Dropbox/UTX-Alex/jan/combined_heights.bed"
temp<-read.table(heightFile)
data<-as.matrix(temp[,4:dim(temp)[2]])
dataLocs<-list(list("eryt", 1, "top"),
              list("tall", 1,"bottom"),
              list("ecfc",3,"top"),
              list("other",c(3,5),c("top","top")),
              list("meka",c(3,7),c("top","top")),
              list("hspc",c(3,7),c("top","bottom")),
              list("diff", 7, "top"))
normData<-qn(data)
prc<-prcomp(t(normData))$rotation
projection<-t(normData)\%*\%prc
weights<-prc[,1]
countNMers(fastaFile,weights,4)
}

